Integrate the Bestof extension
The <bestof> tag shows Wigo votes with a certain prefix ordered by their
vote count. Previously it was in a separate extension, implemented by
parsing the article text with some ad-hoc regexes in an
ArticleSaveComplete hook. I have instead captured the data from the
regular parser and used a secondary update to write it to the database.

Change-Id: I51eba98538d33ca4cbaa3f9c8e9abdb14024bbe2