Remove last remnants of $wgAjaxExportList
This means that eventually this extension will be compatible with MW 1.38+.
That'll probably require a few "use" statements here and there as well as amending the existing "use" statements.

The JS file diff is ugly since I did end up renaming like all the variables to lowerCamelCase, as per the coding conventions, and I also applied some other syntax sugar there.

This is *very* lightly tested with a simple poll, like this:
<vote poll=foobar>I am '''testing''', <u>look!</u></vote>

And by voting up, down and "revoke vote"/neutral on this.
Everything seemed to be working as you'd expect, but again, emphasis on the "lightly" part more than on the "tested" part.

Bug: T42787
Change-Id: I9fd6bc23760821d576ef34647df94f05e0434e6f