Low-effort patch to make Wigo3 (also) work on MW 1.43(+)
The MultiParserHooks.php file was throwing an E_DEPRECATED (on 1.38+), but it'd at least still work as intended.
The API file was throwing "Wikimedia\Rdbms\DBLanguageError: The unique key array should contain a single unique index", which was causing the query to fail.

Disclaimer: I've not tested the API change on 1.35 but I think it should still work.
There are still some weird errors/glitches on 1.43, such as your prior vote not apparently being properly highlighted (whereas that feature works just fine on 1.35), but at least this patch gets some of the core functionality of Wigo3 working on a supported LTS MW version.

I've also not tested the legacy <multi>/<slider> tags, and whatever testing I did on this some weeks ago was fairly quick nevertheless.

Change-Id: Ia424a7d14a0a3e32e4eee42b949151e8e0a68092